continue webpack refactoring
