+++ .jshintrc
@@ -51,24 +51,14 @@
     "globals": {
         "jQuery": true,
         "$": true,
-        "moment": true,
         "pjax_config_page": true,
         "page": true,
         "require": true,
         "module": true,
         "TUser": true,
         "BinarySocket": true,
-        "Portfolio": true,
-        "PortfolioWS": true,
         "addComma": true,
-        "addTooltip": true,
-        "buildOauthApps": true,
-        "console": true,
-        "format_money": true,
-        "japanese_client": true,
-        "showLoadingImage": true,
-        "showTooltip": true,
-        "toJapanTimeIfNeeded": true
+        "console": true
     }
 
     // TODO: Options to enable later:
