+++ src/javascript/binary/websocket_pages/socket.js
@@ -1,25 +1,27 @@
-var getSocketURL = require('../../config').getSocketURL;
-var getAppId = require('../../config').getAppId;
-var Login = require('../base/login').Login;
-var objectNotEmpty = require('../base/utility').objectNotEmpty;
-var CommonFunctions = require('../common_functions/common_functions').CommonFunctions;
-var AccountOpening = require('../common_functions/account_opening').AccountOpening;
-var SessionDurationLimit = require('../common_functions/session_duration_limit').SessionDurationLimit;
-var checkClientsCountry = require('../common_functions/country_base').checkClientsCountry;
-var Cashier = require('../websocket_pages/cashier/cashier').Cashier;
-var PaymentAgentWithdrawWS = require('../websocket_pages/cashier/payment_agent_withdrawws').PaymentAgentWithdrawWS;
+var getSocketURL              = require('../../config').getSocketURL;
+var getAppId                  = require('../../config').getAppId;
+var Login                     = require('../base/login').Login;
+var objectNotEmpty            = require('../base/utility').objectNotEmpty;
+var CommonFunctions           = require('../common_functions/common_functions').CommonFunctions;
+var AccountOpening            = require('../common_functions/account_opening').AccountOpening;
+var SessionDurationLimit      = require('../common_functions/session_duration_limit').SessionDurationLimit;
+var checkClientsCountry       = require('../common_functions/country_base').checkClientsCountry;
+var Cashier                   = require('./cashier/cashier').Cashier;
+var PaymentAgentWithdrawWS    = require('./cashier/payment_agent_withdrawws').PaymentAgentWithdrawWS;
 var create_language_drop_down = require('../common_functions/attach_dom/language_dropdown').create_language_drop_down;
-var TNCApproval = require('../websocket_pages/user/tnc_approval').TNCApproval;
-var ViewPopupWS = require('../websocket_pages/user/view_popup/view_popupws').ViewPopupWS;
-var ViewBalanceUI = require('../websocket_pages/user/viewbalance/viewbalance.ui').ViewBalanceUI;
-var Cookies = require('../../lib/js-cookie');
-var State = require('../base/storage').State;
-var Highchart     = require('./trade/charts/highchartws').Highchart;
-var WSTickDisplay = require('./trade/tick_trade').WSTickDisplay;
-var TradePage      = require('./trade/tradepage').TradePage;
-var Notifications  = require('./trade/notifications').Notifications;
-var TradePage_Beta = require('./trade/beta/tradepage').TradePage_Beta;
-var MBTradePage    = require('./mb_trade/mb_tradepage').MBTradePage;
+var TNCApproval               = require('./user/tnc_approval').TNCApproval;
+var ViewPopupWS               = require('./user/view_popup/view_popupws').ViewPopupWS;
+var ViewBalanceUI             = require('./user/viewbalance/viewbalance.ui').ViewBalanceUI;
+var Cookies                   = require('../../lib/js-cookie');
+var State                     = require('../base/storage').State;
+var Highchart                 = require('./trade/charts/highchartws').Highchart;
+var WSTickDisplay             = require('./trade/tick_trade').WSTickDisplay;
+var TradePage                 = require('./trade/tradepage').TradePage;
+var Notifications             = require('./trade/notifications').Notifications;
+var TradePage_Beta            = require('./trade/beta/tradepage').TradePage_Beta;
+var MBTradePage               = require('./mb_trade/mb_tradepage').MBTradePage;
+var RealityCheck              = require('./user/reality_check/reality_check.init').RealityCheck;
+var RealityCheckData          = require('./user/reality_check/reality_check.data').RealityCheckData;
 
 /*
  * It provides a abstraction layer over native javascript Websocket.
